added Order and Status tables

+++ furniture_store/src/main/java/com/ltp/furniture_store/repository/PermissionTypeRepository.java
@@ -3,5 +3,5 @@ package com.ltp.furniture_store.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.ltp.furniture_store.entity.PermissionType;
-public interface PermissionTypeRepository extends JpaRepository<PermissionType, Long> {
+public interface PermissionTypeRepository extends JpaRepository<PermissionType, Short> {
 }
